Refactor BlockBinder into partials
