+++ src/Raven.CodeAnalysis/Binder/BlockBinder.cs
@@ -422,500 +422,6 @@ partial class BlockBinder : Binder
         return boundNode;
     }
 
-    private BoundStatement BindExpressionStatement(ExpressionStatementSyntax expressionStmt)
-    {
-        var expr = BindExpression(expressionStmt.Expression, allowReturn: true);
-
-        if (expr is BoundMethodGroupExpression methodGroup && methodGroup.GetConvertedType() is null)
-        {
-            expr = ReportMethodGroupRequiresDelegate(methodGroup, expressionStmt.Expression);
-            CacheBoundNode(expressionStmt.Expression, expr);
-        }
-
-        return ExpressionToStatement(expr);
-    }
-
-    private BoundStatement BindIfStatement(IfStatementSyntax ifStmt)
-    {
-        var condition = BindExpression(ifStmt.Condition);
-        var thenBound = BindStatement(ifStmt.ThenStatement);
-        BoundStatement? elseBound = null;
-        if (ifStmt.ElseStatement is not null)
-            elseBound = BindStatement(ifStmt.ElseStatement);
-        return new BoundIfStatement(condition, thenBound, elseBound);
-    }
-
-    private BoundStatement BindWhileStatement(WhileStatementSyntax whileStmt)
-    {
-        var condition = BindExpression(whileStmt.Condition);
-
-        var body = BindStatementInLoop(whileStmt.Statement);
-
-        return new BoundWhileStatement(condition, body);
-    }
-
-    private BoundStatement BindTryStatement(TryStatementSyntax tryStmt)
-    {
-        var tryBlock = BindBlockStatement(tryStmt.Block);
-
-        var catchBuilder = ImmutableArray.CreateBuilder<BoundCatchClause>();
-        foreach (var catchClause in tryStmt.CatchClauses)
-        {
-            catchBuilder.Add(BindCatchClause(catchClause));
-        }
-
-        BoundBlockStatement? finallyBlock = null;
-        if (tryStmt.FinallyClause is { } finallyClause)
-            finallyBlock = BindBlockStatement(finallyClause.Block);
-
-        if (catchBuilder.Count == 0 && finallyBlock is null)
-            return tryBlock;
-
-        return new BoundTryStatement(tryBlock, catchBuilder.ToImmutable(), finallyBlock);
-    }
-
-    private BoundStatement BindForStatement(ForStatementSyntax forStmt)
-    {
-        var loopBinder = (BlockBinder)SemanticModel.GetBinder(forStmt, this)!;
-
-        var collection = BindExpression(forStmt.Expression);
-
-        ITypeSymbol elementType = collection.Type is IArrayTypeSymbol array
-            ? array.ElementType
-            : Compilation.GetSpecialType(SpecialType.System_Object);
-
-        var local = loopBinder.CreateLocalSymbol(forStmt, forStmt.Identifier.ValueText, isMutable: false, elementType);
-
-        var body = loopBinder.BindStatementInLoop(forStmt.Body);
-
-        return new BoundForStatement(local, collection, body);
-    }
-
-    private BoundCatchClause BindCatchClause(CatchClauseSyntax catchClause)
-    {
-        var exceptionBase = Compilation.GetTypeByMetadataName("System.Exception") ?? Compilation.ErrorTypeSymbol;
-        var exceptionType = exceptionBase;
-
-        SourceLocalSymbol? localSymbol = null;
-
-        if (catchClause.Declaration is { } declaration)
-        {
-            var declaredType = ResolveType(declaration.Type);
-            exceptionType = declaredType;
-
-            if (exceptionBase.TypeKind != TypeKind.Error &&
-                declaredType.TypeKind != TypeKind.Error &&
-                !IsAssignable(exceptionBase, declaredType, out _))
-            {
-                _diagnostics.ReportCatchTypeMustDeriveFromSystemException(
-                    declaredType.ToDisplayStringKeywordAware(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                    declaration.Type.GetLocation());
-            }
-
-            if (declaration.Identifier is { } identifier &&
-                !identifier.IsMissing &&
-                identifier.Kind == SyntaxKind.IdentifierToken)
-            {
-                var name = identifier.Text;
-
-                if (LookupSymbol(name) is ILocalSymbol or IParameterSymbol or IFieldSymbol)
-                    _diagnostics.ReportVariableShadowsOuterScope(name, identifier.GetLocation());
-
-                _scopeDepth++;
-                localSymbol = CreateLocalSymbol(declaration, name, isMutable: false, declaredType);
-                _scopeDepth--;
-            }
-        }
-
-        var block = BindBlockStatement(catchClause.Block);
-
-        return new BoundCatchClause(exceptionType, localSymbol, block);
-    }
-
-    private BoundStatement ExpressionToStatement(BoundExpression expression)
-    {
-        return expression switch
-        {
-            BoundIfExpression ifExpr => new BoundIfStatement(ifExpr.Condition,
-                ExpressionToStatement(ifExpr.ThenBranch),
-                ifExpr.ElseBranch is not null ? ExpressionToStatement(ifExpr.ElseBranch) : null),
-            BoundBlockExpression blockExpr => new BoundBlockStatement(blockExpr.Statements, blockExpr.LocalsToDispose),
-            BoundAssignmentExpression assignmentExpr => new BoundAssignmentStatement(assignmentExpr),
-            _ => new BoundExpressionStatement(expression),
-        };
-    }
-
-    private BoundStatement BindReturnStatement(ReturnStatementSyntax returnStatement)
-    {
-        BoundExpression? expr = null;
-
-        if (returnStatement.Expression is not null)
-            expr = BindExpression(returnStatement.Expression);
-
-        if (_containingSymbol is IMethodSymbol method)
-        {
-            if (expr is null)
-            {
-                var unit = Compilation.GetSpecialType(SpecialType.System_Unit);
-                if (!IsAssignable(method.ReturnType, unit, out _))
-                    _diagnostics.ReportCannotConvertFromTypeToType(
-                        unit.ToDisplayStringKeywordAware(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                        method.ReturnType.ToDisplayStringKeywordAware(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                        returnStatement.GetLocation());
-            }
-            else if (expr is not null &&
-                     ShouldAttemptConversion(expr) &&
-                     method.ReturnType.TypeKind != TypeKind.Error)
-            {
-                if (!IsAssignable(method.ReturnType, expr.Type, out var conversion))
-                {
-                    _diagnostics.ReportCannotConvertFromTypeToType(
-                        expr.Type.ToDisplayStringKeywordAware(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                        method.ReturnType.ToDisplayStringKeywordAware(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                        returnStatement.Expression!.GetLocation());
-                }
-                else
-                {
-                    expr = ApplyConversion(expr, method.ReturnType, conversion, returnStatement.Expression!);
-                }
-            }
-        }
-
-        return new BoundReturnStatement(expr);
-    }
-
-    public Dictionary<string, IMethodSymbol> _functions = new();
-
-    protected static bool HaveSameSignature(IMethodSymbol first, IMethodSymbol second)
-    {
-        if (first.Parameters.Length != second.Parameters.Length)
-            return false;
-
-        for (int i = 0; i < first.Parameters.Length; i++)
-        {
-            if (!SymbolEqualityComparer.Default.Equals(first.Parameters[i].Type, second.Parameters[i].Type))
-                return false;
-        }
-
-        return true;
-    }
-
-    public virtual BoundBlockStatement BindBlockStatement(BlockStatementSyntax block)
-    {
-        if (TryGetCachedBoundNode(block) is BoundStatement cached)
-            return (BoundBlockStatement)cached;
-
-        _scopeDepth++;
-        var depth = _scopeDepth;
-
-        EnsureLabelsDeclared(block);
-
-        foreach (var stmt in block.Statements)
-        {
-            if (stmt is FunctionStatementSyntax func)
-            {
-                var functionBinder = SemanticModel.GetBinder(func, this);
-                if (functionBinder is FunctionBinder lfBinder)
-                {
-                    var symbol = lfBinder.GetMethodSymbol();
-                    if (_functions.TryGetValue(symbol.Name, out var existing) && HaveSameSignature(existing, symbol))
-                        _diagnostics.ReportFunctionAlreadyDefined(symbol.Name, func.Identifier.GetLocation());
-                    else
-                        _functions[symbol.Name] = symbol;
-                }
-            }
-        }
-
-        var boundStatements = new List<BoundStatement>(block.Statements.Count);
-        foreach (var stmt in block.Statements)
-        {
-            var bound = BindStatement(stmt);
-            boundStatements.Add(bound);
-        }
-
-        var localsAtDepth = _localsToDispose
-            .Where(l => l.Depth == depth)
-            .Select(l => l.Local)
-            .ToList();
-
-        if (localsAtDepth.Count > 0)
-            _localsToDispose.RemoveAll(l => l.Depth == depth);
-
-        var blockStmt = new BoundBlockStatement(boundStatements.ToArray(), localsAtDepth.ToImmutableArray());
-        CacheBoundNode(block, blockStmt);
-
-        foreach (var name in _locals.Where(kvp => kvp.Value.Depth == depth).Select(kvp => kvp.Key).ToList())
-            _locals.Remove(name);
-
-        _scopeDepth--;
-        return blockStmt;
-    }
-
-    public virtual BoundBlockExpression BindBlock(BlockSyntax block, bool allowReturn = true)
-    {
-        if (TryGetCachedBoundNode(block) is BoundExpression cached)
-            return (BoundBlockExpression)cached;
-
-        _scopeDepth++;
-        var depth = _scopeDepth;
-
-        if (!allowReturn)
-            _expressionContextDepth++;
-
-        EnsureLabelsDeclared(block);
-
-        // Step 1: Pre-declare all functions
-        foreach (var stmt in block.Statements)
-        {
-            if (stmt is FunctionStatementSyntax func)
-            {
-                var functionBinder = SemanticModel.GetBinder(func, this);
-                if (functionBinder is FunctionBinder lfBinder)
-                {
-                    var symbol = lfBinder.GetMethodSymbol();
-                    if (_functions.TryGetValue(symbol.Name, out var existing) && HaveSameSignature(existing, symbol))
-                        _diagnostics.ReportFunctionAlreadyDefined(symbol.Name, func.Identifier.GetLocation());
-                    else
-                        _functions[symbol.Name] = symbol;
-                }
-            }
-        }
-
-        // Step 2: Bind and cache all statements
-        var boundStatements = new List<BoundStatement>(block.Statements.Count);
-        foreach (var stmt in block.Statements)
-        {
-            BoundStatement bound;
-            if (!allowReturn && stmt is ReturnStatementSyntax ret)
-            {
-                _diagnostics.ReportReturnStatementInExpression(stmt.GetLocation());
-                var expr = ret.Expression is null
-                    ? new BoundUnitExpression(Compilation.GetSpecialType(SpecialType.System_Unit))
-                    : BindExpression(ret.Expression);
-                bound = new BoundExpressionStatement(expr);
-            }
-            else
-            {
-                bound = BindStatement(stmt);
-                if (!allowReturn && bound is BoundReturnStatement br)
-                {
-                    _diagnostics.ReportReturnStatementInExpression(stmt.GetLocation());
-                    var expr = br.Expression ?? new BoundUnitExpression(Compilation.GetSpecialType(SpecialType.System_Unit));
-                    bound = new BoundExpressionStatement(expr);
-                }
-            }
-            boundStatements.Add(bound);
-        }
-
-        // Step 3: Create and cache the block
-        var unitType = Compilation.GetSpecialType(SpecialType.System_Unit);
-        var localsAtDepth = _localsToDispose
-            .Where(l => l.Depth == depth)
-            .Select(l => l.Local)
-            .ToList();
-
-        if (localsAtDepth.Count > 0)
-            _localsToDispose.RemoveAll(l => l.Depth == depth);
-
-        var blockExpr = new BoundBlockExpression(boundStatements.ToArray(), unitType, localsAtDepth.ToImmutableArray());
-        CacheBoundNode(block, blockExpr);
-
-        foreach (var name in _locals.Where(kvp => kvp.Value.Depth == depth).Select(kvp => kvp.Key).ToList())
-            _locals.Remove(name);
-
-        _scopeDepth--;
-        if (!allowReturn)
-            _expressionContextDepth--;
-        return blockExpr;
-    }
-
-    private void EnsureLabelsDeclared(SyntaxNode node)
-    {
-        if (!_labelDeclarationNodes.Add(node))
-            return;
-
-        var stack = new Stack<SyntaxNode>();
-        stack.Push(node);
-
-        while (stack.Count > 0)
-        {
-            var current = stack.Pop();
-
-            if (current is FunctionStatementSyntax)
-                continue;
-
-            if (current is LabeledStatementSyntax labeled)
-            {
-                _ = DeclareLabelSymbol(labeled);
-                stack.Push(labeled.Statement);
-                continue;
-            }
-
-            foreach (var child in current.ChildNodes())
-                stack.Push(child);
-        }
-    }
-
-    private ILabelSymbol DeclareLabelSymbol(LabeledStatementSyntax labeledStatement)
-    {
-        if (_labelsBySyntax.TryGetValue(labeledStatement, out var existing))
-            return existing;
-
-        var identifier = labeledStatement.Identifier;
-        if (identifier.IsMissing)
-            return CreateLabelSymbol(string.Empty, identifier.GetLocation(), labeledStatement.GetReference());
-
-        if (SyntaxFacts.IsReservedWordKind(identifier.Kind))
-        {
-            var identifierName = identifier.ValueText;
-            _diagnostics.ReportReservedWordCannotBeLabel(identifierName, identifier.GetLocation());
-            return CreateLabelSymbol(string.Empty, identifier.GetLocation(), labeledStatement.GetReference());
-        }
-
-        var name = identifier.ValueText;
-
-        if (_labelsByName.TryGetValue(name, out var conflict))
-        {
-            _diagnostics.ReportLabelAlreadyDefined(name, identifier.GetLocation());
-            return conflict;
-        }
-
-        var symbol = CreateLabelSymbol(name, identifier.GetLocation(), labeledStatement.GetReference());
-
-        _labelsByName[name] = symbol;
-        _labelsBySyntax[labeledStatement] = symbol;
-        _syntaxByLabel[symbol] = labeledStatement;
-        SemanticModel?.RegisterLabel(labeledStatement, symbol);
-
-        return symbol;
-    }
-
-    private BoundStatement BindLabeledStatement(LabeledStatementSyntax labeledStatement)
-    {
-        if (_expressionContextDepth > 0)
-            _diagnostics.ReportLabelInExpression(labeledStatement.Identifier.GetLocation());
-
-        var labelSymbol = DeclareLabelSymbol(labeledStatement);
-        var boundStatement = BindStatement(labeledStatement.Statement);
-        var bound = new BoundLabeledStatement(labelSymbol, boundStatement);
-        CacheBoundNode(labeledStatement, bound);
-        return bound;
-    }
-
-    private BoundStatement BindGotoStatement(GotoStatementSyntax gotoStatement)
-    {
-        if (_expressionContextDepth > 0)
-            _diagnostics.ReportGotoStatementInExpression(gotoStatement.GotoKeyword.GetLocation());
-
-        var identifier = gotoStatement.Identifier;
-        if (identifier.IsMissing)
-        {
-            var errorSymbol = CreateLabelSymbol(string.Empty, identifier.GetLocation(), gotoStatement.GetReference());
-            var boundError = new BoundGotoStatement(errorSymbol);
-            CacheBoundNode(gotoStatement, boundError);
-            return boundError;
-        }
-
-        if (SyntaxFacts.IsReservedWordKind(identifier.Kind))
-        {
-            var identifierName = identifier.ValueText;
-            _diagnostics.ReportReservedWordCannotBeLabel(identifierName, identifier.GetLocation());
-            var errorSymbol = CreateLabelSymbol(string.Empty, identifier.GetLocation(), gotoStatement.GetReference());
-            var boundError = new BoundGotoStatement(errorSymbol);
-            CacheBoundNode(gotoStatement, boundError);
-            return boundError;
-        }
-
-        var name = identifier.ValueText;
-        if (!_labelsByName.TryGetValue(name, out var label))
-        {
-            _diagnostics.ReportLabelNotFound(name, identifier.GetLocation());
-            label = CreateLabelSymbol(name, identifier.GetLocation(), gotoStatement.GetReference());
-        }
-
-        var isBackward = false;
-        if (_syntaxByLabel.TryGetValue(label, out var labeledSyntax))
-            isBackward = labeledSyntax.Span.Start < gotoStatement.Span.Start;
-
-        SemanticModel?.RegisterGoto(gotoStatement, label);
-
-        var bound = new BoundGotoStatement(label, isBackward);
-        CacheBoundNode(gotoStatement, bound);
-        return bound;
-    }
-
-    private BoundStatement BindBreakStatement(BreakStatementSyntax breakStatement)
-    {
-        var location = breakStatement.BreakKeyword.GetLocation();
-
-        if (_expressionContextDepth > 0)
-        {
-            _diagnostics.ReportBreakStatementInExpression(location);
-        }
-        else if (_loopDepth == 0)
-        {
-            _diagnostics.ReportBreakStatementNotWithinLoop(location);
-        }
-
-        var bound = new BoundBreakStatement();
-        CacheBoundNode(breakStatement, bound);
-        return bound;
-    }
-
-    private BoundStatement BindContinueStatement(ContinueStatementSyntax continueStatement)
-    {
-        var location = continueStatement.ContinueKeyword.GetLocation();
-
-        if (_expressionContextDepth > 0)
-        {
-            _diagnostics.ReportContinueStatementInExpression(location);
-        }
-        else if (_loopDepth == 0)
-        {
-            _diagnostics.ReportContinueStatementNotWithinLoop(location);
-        }
-
-        var bound = new BoundContinueStatement();
-        CacheBoundNode(continueStatement, bound);
-        return bound;
-    }
-
-    public BoundStatement BindStatementInLoop(StatementSyntax syntax)
-    {
-        var previous = EnterLoop();
-        try
-        {
-            return BindStatement(syntax);
-        }
-        finally
-        {
-            ExitLoop(previous);
-        }
-    }
-
-    private int EnterLoop()
-    {
-        var previous = _loopDepth;
-        _loopDepth++;
-        return previous;
-    }
-
-    private void ExitLoop(int previous)
-    {
-        _loopDepth = previous;
-    }
-
-    private ILabelSymbol CreateLabelSymbol(string name, Location location, SyntaxReference reference)
-    {
-        return new LabelSymbol(
-            name,
-            _containingSymbol,
-            _containingSymbol.ContainingType as INamedTypeSymbol,
-            _containingSymbol?.ContainingNamespace,
-            [location],
-            [reference]);
-    }
-
     public BoundExpression BindExpression(ExpressionSyntax syntax, bool allowReturn)
     {
         var previous = _allowReturnsInExpression;
@@ -1067,232 +573,6 @@ partial class BlockBinder : Binder
 
         //_diagnostics.ReportInvalidAddressOf(syntax.Expression.GetLocation());
         return new BoundErrorExpression(Compilation.ErrorTypeSymbol, null, BoundExpressionReason.ArgumentBindingFailed /*,.InvalidAddressOfTarget */);
-    }
-
-    private BoundExpression BindLambdaExpression(LambdaExpressionSyntax syntax)
-    {
-        // 1. Extract parameter syntax
-        var parameterSyntaxes = syntax switch
-        {
-            SimpleLambdaExpressionSyntax s => new[] { s.Parameter },
-            ParenthesizedLambdaExpressionSyntax p => p.ParameterList.Parameters.ToArray(),
-            _ => throw new NotSupportedException("Unknown lambda syntax")
-        };
-
-        var targetType = GetTargetType(syntax);
-        var candidateDelegates = GetLambdaDelegateTargets(syntax);
-        if (candidateDelegates.IsDefaultOrEmpty)
-            candidateDelegates = ComputeLambdaDelegateTargets(syntax);
-        var suppressedDiagnostics = ImmutableArray.CreateBuilder<SuppressedLambdaDiagnostic>();
-
-        INamedTypeSymbol? targetDelegate = targetType as INamedTypeSymbol;
-        if (targetDelegate?.TypeKind != TypeKind.Delegate)
-            targetDelegate = null;
-
-        if (targetDelegate is not null && candidateDelegates.IsDefault)
-        {
-            candidateDelegates = ImmutableArray.Create(targetDelegate);
-        }
-        else if (targetDelegate is not null && !candidateDelegates.Contains(targetDelegate, SymbolEqualityComparer.Default))
-        {
-            candidateDelegates = candidateDelegates.Add(targetDelegate);
-        }
-
-        if (candidateDelegates.IsDefault)
-            candidateDelegates = ImmutableArray<INamedTypeSymbol>.Empty;
-
-        INamedTypeSymbol? primaryDelegate = targetDelegate ?? candidateDelegates.FirstOrDefault();
-        var targetSignature = primaryDelegate?.GetDelegateInvokeMethod();
-
-        // 2. Create parameter symbols
-        var parameterSymbols = new List<IParameterSymbol>();
-        for (int index = 0; index < parameterSyntaxes.Length; index++)
-        {
-            var parameterSyntax = parameterSyntaxes[index];
-            var annotation = parameterSyntax.TypeAnnotation;
-            var typeSyntax = annotation?.Type;
-            var refKind = RefKind.None;
-
-            if (typeSyntax is ByRefTypeSyntax byRefSyntax)
-            {
-                refKind = parameterSyntax.Modifiers.Any(m => m.Kind == SyntaxKind.OutKeyword) ? RefKind.Out : RefKind.Ref;
-                typeSyntax = byRefSyntax.ElementType;
-            }
-
-            var targetParam = targetSignature is { } invoke && invoke.Parameters.Length > index
-                ? invoke.Parameters[index]
-                : null;
-
-            ITypeSymbol parameterType;
-            if (typeSyntax is not null)
-            {
-                parameterType = ResolveType(typeSyntax);
-            }
-            else if (targetParam is not null)
-            {
-                parameterType = targetParam.Type;
-                if (refKind == RefKind.None)
-                    refKind = targetParam.RefKind;
-            }
-            else
-            {
-                parameterType = TryInferLambdaParameterType(candidateDelegates, index, out var inferredRefKind)
-                    ?? Compilation.ErrorTypeSymbol;
-
-                if (refKind == RefKind.None && inferredRefKind is not null)
-                    refKind = inferredRefKind.Value;
-
-                if (parameterType.TypeKind == TypeKind.Error)
-                {
-                    var parameterName = parameterSyntax.Identifier.ValueText;
-                    var parameterLocation = parameterSyntax.Identifier.GetLocation();
-
-                    if (candidateDelegates.IsDefaultOrEmpty)
-                    {
-                        _diagnostics.ReportLambdaParameterTypeCannotBeInferred(
-                            parameterName,
-                            parameterLocation);
-                    }
-                    else
-                    {
-                        suppressedDiagnostics.Add(new SuppressedLambdaDiagnostic(parameterName, parameterLocation));
-                    }
-                }
-            }
-
-            var symbol = new SourceParameterSymbol(
-                parameterSyntax.Identifier.ValueText,
-                parameterType,
-                _containingSymbol,
-                _containingSymbol.ContainingType as INamedTypeSymbol,
-                _containingSymbol.ContainingNamespace,
-                [parameterSyntax.GetLocation()],
-                [parameterSyntax.GetReference()],
-                refKind
-            );
-
-            parameterSymbols.Add(symbol);
-        }
-
-        // 3. Resolve explicit return type (if any)
-        TypeSyntax? returnTypeSyntax = syntax switch
-        {
-            SimpleLambdaExpressionSyntax s => s.ReturnType.Type,
-            ParenthesizedLambdaExpressionSyntax p => p.ReturnType.Type,
-            _ => null
-        };
-
-        // 4. Create symbol for the lambda
-        var inferredReturnType = returnTypeSyntax is not null
-            ? ResolveType(returnTypeSyntax)
-            : Compilation.ErrorTypeSymbol; // Temporary fallback; real type inferred below
-
-        var lambdaSymbol = new SourceLambdaSymbol(
-            parameterSymbols,
-            inferredReturnType,
-            _containingSymbol,
-            _containingSymbol.ContainingType as INamedTypeSymbol,
-            _containingSymbol.ContainingNamespace,
-            [syntax.GetLocation()],
-            [syntax.GetReference()]);
-
-        // 5. Bind the body using a new binder scope
-        var lambdaBinder = new LambdaBinder(lambdaSymbol, this);
-
-        foreach (var param in parameterSymbols)
-            lambdaBinder.DeclareParameter(param);
-
-        var bodyExpr = lambdaBinder.BindExpression(syntax.ExpressionBody, allowReturn: true);
-
-        var inferred = bodyExpr.Type ?? ReturnTypeCollector.Infer(bodyExpr);
-
-        ITypeSymbol returnType;
-        if (returnTypeSyntax is not null)
-        {
-            returnType = inferredReturnType;
-            if (inferred is not null &&
-                inferred.TypeKind != TypeKind.Error &&
-                returnType.TypeKind != TypeKind.Error)
-            {
-                if (!IsAssignable(returnType, inferred, out var conversion))
-                {
-                    _diagnostics.ReportCannotConvertFromTypeToType(
-                        inferred.ToDisplayStringKeywordAware(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                        returnType.ToDisplayStringKeywordAware(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                        syntax.ExpressionBody.GetLocation());
-                }
-                else
-                {
-                    bodyExpr = ApplyConversion(bodyExpr, returnType, conversion, syntax.ExpressionBody);
-                }
-            }
-        }
-        else if (targetSignature is { ReturnType: { } targetReturn } && targetReturn.TypeKind != TypeKind.Error)
-        {
-            returnType = targetReturn;
-            if (inferred is not null &&
-                inferred.TypeKind != TypeKind.Error &&
-                returnType.TypeKind != TypeKind.Error)
-            {
-                if (!IsAssignable(returnType, inferred, out var conversion))
-                {
-                    _diagnostics.ReportCannotConvertFromTypeToType(
-                        inferred.ToDisplayStringKeywordAware(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                        returnType.ToDisplayStringKeywordAware(SymbolDisplayFormat.MinimallyQualifiedFormat),
-                        syntax.ExpressionBody.GetLocation());
-                }
-                else
-                {
-                    bodyExpr = ApplyConversion(bodyExpr, returnType, conversion, syntax.ExpressionBody);
-                }
-            }
-        }
-        else
-        {
-            returnType = inferred ?? Compilation.ErrorTypeSymbol;
-        }
-
-        lambdaBinder.SetLambdaBody(bodyExpr);
-
-        var capturedVariables = lambdaBinder.AnalyzeCapturedVariables();
-
-        if (lambdaSymbol is SourceLambdaSymbol sourceLambdaSymbol)
-            sourceLambdaSymbol.SetCapturedVariables(capturedVariables);
-
-        // 7. Construct delegate type (e.g., Func<...> or custom delegate)
-        var delegateType = primaryDelegate is not null &&
-            targetSignature is not null &&
-            targetSignature.Parameters.Length == parameterSymbols.Count &&
-            SymbolEqualityComparer.Default.Equals(returnType, targetSignature.ReturnType) &&
-            parameterSymbols
-                .Zip(targetSignature.Parameters, (parameter, target) =>
-                    SymbolEqualityComparer.Default.Equals(parameter.Type, target.Type) && parameter.RefKind == target.RefKind)
-                .All(match => match)
-            ? primaryDelegate
-            : Compilation.CreateFunctionTypeSymbol(
-                parameterSymbols.Select(p => p.Type).ToArray(),
-                returnType
-            );
-
-        // 7. Update lambda symbol if needed
-        if (lambdaSymbol is SourceLambdaSymbol mutable)
-        {
-            mutable.SetReturnType(returnType);
-            mutable.SetDelegateType(delegateType);
-        }
-
-        // 8. Return a fully bound lambda expression
-        var boundLambda = new BoundLambdaExpression(parameterSymbols, returnType, bodyExpr, lambdaSymbol, delegateType, capturedVariables, candidateDelegates);
-
-        var suppressed = suppressedDiagnostics.ToImmutable();
-        if ((!candidateDelegates.IsDefaultOrEmpty || suppressed.Length > 0) && lambdaSymbol is SourceLambdaSymbol source)
-        {
-            var parameters = parameterSymbols.ToImmutableArray();
-            var unbound = new BoundUnboundLambda(source, syntax, parameters, candidateDelegates, suppressed);
-            boundLambda.AttachUnbound(unbound);
-        }
-
-        return boundLambda;
     }
 
     private BoundExpression BindMissingExpression(ExpressionSyntax.Missing missing)
@@ -2382,204 +1662,6 @@ partial class BlockBinder : Binder
         return null;
     }
 
-    private ImmutableArray<INamedTypeSymbol> GetLambdaDelegateTargets(LambdaExpressionSyntax syntax)
-    {
-        if (_lambdaDelegateTargets.TryGetValue(syntax, out var targets))
-            return targets;
-
-        return ImmutableArray<INamedTypeSymbol>.Empty;
-    }
-
-    private ImmutableArray<INamedTypeSymbol> ComputeLambdaDelegateTargets(LambdaExpressionSyntax syntax)
-    {
-        if (syntax.Parent is not ArgumentSyntax argument ||
-            argument.Parent is not ArgumentListSyntax argumentList)
-        {
-            return ImmutableArray<INamedTypeSymbol>.Empty;
-        }
-
-        var invocation = argumentList.Parent as InvocationExpressionSyntax;
-        if (invocation is null)
-            return ImmutableArray<INamedTypeSymbol>.Empty;
-
-        var argumentExpression = argument.Expression;
-        var parameterIndex = 0;
-        foreach (var candidate in argumentList.Arguments)
-        {
-            if (candidate == argument)
-                break;
-            parameterIndex++;
-        }
-
-        ImmutableArray<IMethodSymbol> methods = default;
-        var extensionReceiverImplicit = false;
-
-        switch (invocation.Expression)
-        {
-            case MemberAccessExpressionSyntax memberAccess:
-                {
-                    var boundMember = BindMemberAccessExpression(memberAccess);
-                    if (boundMember is BoundMethodGroupExpression methodGroup)
-                    {
-                        extensionReceiverImplicit = methodGroup.Receiver is not null && IsExtensionReceiver(methodGroup.Receiver);
-                        methods = FilterMethodsForLambda(methodGroup.Methods, parameterIndex, argumentExpression, extensionReceiverImplicit);
-                    }
-                    else if (boundMember is BoundMemberAccessExpression { Receiver: var receiver, Member: IMethodSymbol method })
-                    {
-                        extensionReceiverImplicit = receiver is not null && IsExtensionReceiver(receiver);
-                        methods = ImmutableArray.Create(method);
-                    }
-
-                    break;
-                }
-
-            case MemberBindingExpressionSyntax memberBinding:
-                {
-                    var boundMember = BindMemberBindingExpression(memberBinding);
-                    if (boundMember is BoundMethodGroupExpression methodGroup)
-                    {
-                        extensionReceiverImplicit = methodGroup.Receiver is not null && IsExtensionReceiver(methodGroup.Receiver);
-                        methods = FilterMethodsForLambda(methodGroup.Methods, parameterIndex, argumentExpression, extensionReceiverImplicit);
-                    }
-                    else if (boundMember is BoundMemberAccessExpression { Receiver: var receiver, Member: IMethodSymbol method })
-                    {
-                        extensionReceiverImplicit = receiver is not null && IsExtensionReceiver(receiver);
-                        methods = ImmutableArray.Create(method);
-                    }
-
-                    break;
-                }
-
-            case IdentifierNameSyntax identifier:
-                {
-                    var candidates = new SymbolQuery(identifier.Identifier.ValueText)
-                        .LookupMethods(this)
-                        .ToImmutableArray();
-
-                    var accessible = GetAccessibleMethods(candidates, identifier.Identifier.GetLocation(), reportIfInaccessible: false);
-                    if (!accessible.IsDefaultOrEmpty)
-                    {
-                        methods = FilterMethodsForLambda(accessible, parameterIndex, argumentExpression, extensionReceiverImplicit: false);
-                    }
-
-                    break;
-                }
-        }
-
-        if (methods.IsDefaultOrEmpty)
-            return ImmutableArray<INamedTypeSymbol>.Empty;
-
-        var delegates = ExtractLambdaDelegates(methods, parameterIndex, extensionReceiverImplicit);
-        if (!delegates.IsDefaultOrEmpty)
-            _lambdaDelegateTargets[syntax] = delegates;
-        else
-            _lambdaDelegateTargets.Remove(syntax);
-        return delegates;
-    }
-
-    private ImmutableArray<INamedTypeSymbol> ExtractLambdaDelegates(
-        ImmutableArray<IMethodSymbol> methods,
-        int parameterIndex,
-        bool extensionReceiverImplicit)
-    {
-        if (methods.IsDefaultOrEmpty)
-            return ImmutableArray<INamedTypeSymbol>.Empty;
-
-        var builder = ImmutableArray.CreateBuilder<INamedTypeSymbol>();
-
-        foreach (var method in methods)
-        {
-            if (!TryGetLambdaParameter(method, parameterIndex, extensionReceiverImplicit, out var parameter))
-                continue;
-
-            if (parameter.Type is INamedTypeSymbol delegateType && delegateType.TypeKind == TypeKind.Delegate)
-            {
-                if (!builder.Any(existing => SymbolEqualityComparer.Default.Equals(existing, delegateType)))
-                    builder.Add(delegateType);
-            }
-        }
-
-        return builder.ToImmutable();
-    }
-
-    private void RecordLambdaTargets(
-        ExpressionSyntax argumentExpression,
-        ImmutableArray<IMethodSymbol> methods,
-        int parameterIndex,
-        bool extensionReceiverImplicit = false)
-    {
-        if (argumentExpression is not LambdaExpressionSyntax lambda)
-        {
-            if (argumentExpression is LambdaExpressionSyntax lambdaExpression)
-                _lambdaDelegateTargets.Remove(lambdaExpression);
-            return;
-        }
-
-        if (methods.IsDefaultOrEmpty)
-            return;
-
-        var builder = ImmutableArray.CreateBuilder<INamedTypeSymbol>();
-        var seen = new HashSet<INamedTypeSymbol>(SymbolEqualityComparer.Default);
-
-        if (_lambdaDelegateTargets.TryGetValue(lambda, out var existing) && !existing.IsDefaultOrEmpty)
-        {
-            foreach (var candidate in existing)
-            {
-                if (seen.Add(candidate))
-                    builder.Add(candidate);
-            }
-        }
-
-        foreach (var method in methods)
-        {
-            if (!TryGetLambdaParameter(method, parameterIndex, extensionReceiverImplicit, out var parameter))
-                continue;
-
-            if (parameter.Type is INamedTypeSymbol delegateType && delegateType.TypeKind == TypeKind.Delegate)
-            {
-                if (seen.Add(delegateType))
-                    builder.Add(delegateType);
-            }
-        }
-
-        if (builder.Count > 0)
-            _lambdaDelegateTargets[lambda] = builder.ToImmutable();
-    }
-
-    private ITypeSymbol? TryInferLambdaParameterType(ImmutableArray<INamedTypeSymbol> candidateDelegates, int parameterIndex, out RefKind? inferredRefKind)
-    {
-        inferredRefKind = null;
-
-        if (candidateDelegates.IsDefaultOrEmpty)
-            return null;
-
-        ITypeSymbol? inferredType = null;
-
-        foreach (var delegateType in candidateDelegates)
-        {
-            var invoke = delegateType.GetDelegateInvokeMethod();
-            if (invoke is null || invoke.Parameters.Length <= parameterIndex)
-                return null;
-
-            var parameter = invoke.Parameters[parameterIndex];
-
-            if (inferredType is null)
-            {
-                inferredType = parameter.Type;
-                inferredRefKind = parameter.RefKind;
-                continue;
-            }
-
-            if (!SymbolEqualityComparer.Default.Equals(inferredType, parameter.Type))
-                return null;
-
-            if (inferredRefKind != parameter.RefKind)
-                return null;
-        }
-
-        return inferredType;
-    }
-
     protected BoundExpression BindTypeSyntax(TypeSyntax syntax)
     {
         if (syntax is NullTypeSyntax)
@@ -3493,108 +2575,6 @@ partial class BlockBinder : Binder
         ReportSuppressedLambdaDiagnostics(boundArguments);
         _diagnostics.ReportNoOverloadForMethod(methodName, boundArguments.Length, syntax.GetLocation());
         return new BoundErrorExpression(Compilation.ErrorTypeSymbol, null, BoundExpressionReason.OverloadResolutionFailed);
-    }
-
-    private static bool TryGetLambdaParameter(
-        IMethodSymbol method,
-        int argumentIndex,
-        bool extensionReceiverImplicit,
-        out IParameterSymbol? parameter)
-    {
-        var parameterIndex = GetLambdaParameterIndex(method, argumentIndex, extensionReceiverImplicit);
-
-        if (parameterIndex < 0 || parameterIndex >= method.Parameters.Length)
-        {
-            parameter = null;
-            return false;
-        }
-
-        parameter = method.Parameters[parameterIndex];
-        return true;
-    }
-
-    private static int GetLambdaParameterIndex(
-        IMethodSymbol method,
-        int argumentIndex,
-        bool extensionReceiverImplicit)
-    {
-        if (method.IsExtensionMethod && extensionReceiverImplicit)
-            return argumentIndex + 1;
-
-        return argumentIndex;
-    }
-
-    private ITypeSymbol? TryGetCommonParameterType(
-        ImmutableArray<IMethodSymbol> candidates,
-        int index,
-        bool extensionReceiverImplicit = false)
-    {
-        if (candidates.IsDefaultOrEmpty)
-            return null;
-
-        ITypeSymbol? parameterType = null;
-
-        foreach (var candidate in candidates)
-        {
-            if (!TryGetLambdaParameter(candidate, index, extensionReceiverImplicit, out var parameter))
-                return null;
-
-            var candidateType = parameter.Type;
-
-            if (parameterType is null)
-            {
-                parameterType = candidateType;
-                continue;
-            }
-
-            if (!SymbolEqualityComparer.Default.Equals(parameterType, candidateType))
-                return null;
-        }
-
-        return parameterType;
-    }
-
-    private ImmutableArray<IMethodSymbol> FilterMethodsForLambda(
-        ImmutableArray<IMethodSymbol> methods,
-        int parameterIndex,
-        ExpressionSyntax argumentExpression,
-        bool extensionReceiverImplicit)
-    {
-        if (methods.IsDefaultOrEmpty)
-            return methods;
-
-        if (argumentExpression is not LambdaExpressionSyntax lambda)
-            return methods;
-
-        int parameterCount = lambda switch
-        {
-            SimpleLambdaExpressionSyntax => 1,
-            ParenthesizedLambdaExpressionSyntax p => p.ParameterList.Parameters.Count,
-            _ => 0
-        };
-
-        var builder = ImmutableArray.CreateBuilder<IMethodSymbol>();
-
-        foreach (var method in methods)
-        {
-            if (!TryGetLambdaParameter(method, parameterIndex, extensionReceiverImplicit, out var parameter))
-                continue;
-
-            var parameterType = parameter.Type;
-            if (parameterType is INamedTypeSymbol delegateType && delegateType.TypeKind == TypeKind.Delegate)
-            {
-                var invoke = delegateType.GetDelegateInvokeMethod();
-                if (invoke is null)
-                    continue;
-
-                if (invoke.Parameters.Length != parameterCount)
-                    continue;
-            }
-
-            builder.Add(method);
-        }
-
-        return builder.Count == methods.Length ? methods : builder.ToImmutable();
     }
 
     private BoundExpression BindConstructorInvocation(
@@ -4585,203 +3565,6 @@ partial class BlockBinder : Binder
         }
 
         return converted;
-    }
-
-    private bool EnsureLambdaCompatible(IParameterSymbol parameter, BoundLambdaExpression lambda)
-    {
-        if (parameter.Type is not INamedTypeSymbol delegateType)
-            return false;
-
-        return ReplayLambda(lambda, delegateType) is not null;
-    }
-
-    private BoundLambdaExpression? ReplayLambda(BoundLambdaExpression lambda, INamedTypeSymbol delegateType)
-    {
-        var instrumentation = Compilation.PerformanceInstrumentation.LambdaReplay;
-        instrumentation.RecordReplayAttempt();
-
-        if (delegateType.TypeKind != TypeKind.Delegate)
-        {
-            instrumentation.RecordReplayFailure();
-            return null;
-        }
-
-        var syntax = GetLambdaSyntax(lambda);
-
-        if (syntax is not null)
-        {
-            var key = new LambdaRebindKey(syntax, delegateType);
-            if (_reboundLambdaCache.TryGetValue(key, out var cached))
-            {
-                instrumentation.RecordCacheHit();
-                instrumentation.RecordReplaySuccess();
-                return cached;
-            }
-
-            instrumentation.RecordCacheMiss();
-        }
-
-        BoundLambdaExpression? rebound;
-        if (lambda.Unbound is { } unbound)
-        {
-            rebound = ReplayLambda(unbound, delegateType);
-            if (rebound is null)
-            {
-                instrumentation.RecordReplayFailure();
-                return null;
-            }
-
-            instrumentation.RecordReplaySuccess();
-        }
-        else if (lambda.IsCompatibleWithDelegate(delegateType, Compilation))
-        {
-            rebound = lambda;
-            instrumentation.RecordReplaySuccess();
-        }
-        else
-        {
-            instrumentation.RecordReplayFailure();
-            return null;
-        }
-
-        if (syntax is not null && rebound is not null)
-            _reboundLambdaCache[new LambdaRebindKey(syntax, delegateType)] = rebound;
-
-        return rebound;
-    }
-
-    private BoundLambdaExpression? ReplayLambda(BoundUnboundLambda unbound, INamedTypeSymbol delegateType)
-    {
-        var instrumentation = Compilation.PerformanceInstrumentation.LambdaReplay;
-        instrumentation.RecordBindingInvocation();
-
-        var invoke = delegateType.GetDelegateInvokeMethod();
-        if (invoke is null)
-        {
-            instrumentation.RecordBindingFailure();
-            return null;
-        }
-
-        var syntax = unbound.Syntax;
-        var parameterSyntaxes = syntax switch
-        {
-            SimpleLambdaExpressionSyntax simple => new[] { simple.Parameter },
-            ParenthesizedLambdaExpressionSyntax parenthesized => parenthesized.ParameterList.Parameters.ToArray(),
-            _ => Array.Empty<ParameterSyntax>()
-        };
-
-        if (invoke.Parameters.Length != parameterSyntaxes.Length)
-        {
-            instrumentation.RecordBindingFailure();
-            return null;
-        }
-
-        var parameterSymbols = new List<IParameterSymbol>(parameterSyntaxes.Length);
-
-        for (int index = 0; index < parameterSyntaxes.Length; index++)
-        {
-            var parameterSyntax = parameterSyntaxes[index];
-            var delegateParameter = invoke.Parameters[index];
-            var parameterSymbol = new SourceParameterSymbol(
-                parameterSyntax.Identifier.ValueText,
-                delegateParameter.Type,
-                _containingSymbol,
-                _containingSymbol.ContainingType as INamedTypeSymbol,
-                _containingSymbol.ContainingNamespace,
-                [parameterSyntax.GetLocation()],
-                [parameterSyntax.GetReference()],
-                delegateParameter.RefKind);
-
-            parameterSymbols.Add(parameterSymbol);
-        }
-
-        var lambdaSymbol = new SourceLambdaSymbol(
-            parameterSymbols,
-            invoke.ReturnType,
-            _containingSymbol,
-            _containingSymbol.ContainingType as INamedTypeSymbol,
-            _containingSymbol.ContainingNamespace,
-            [syntax.GetLocation()],
-            [syntax.GetReference()]);
-
-        var lambdaBinder = new LambdaBinder(lambdaSymbol, this);
-
-        foreach (var parameter in parameterSymbols)
-            lambdaBinder.DeclareParameter(parameter);
-
-        var body = lambdaBinder.BindExpression(syntax.ExpressionBody, allowReturn: true);
-        var inferred = body.Type ?? ReturnTypeCollector.Infer(body);
-        var returnType = invoke.ReturnType;
-
-        if (inferred is not null &&
-            inferred.TypeKind != TypeKind.Error &&
-            returnType.TypeKind != TypeKind.Error)
-        {
-            if (!IsAssignable(returnType, inferred, out var conversion))
-            {
-                instrumentation.RecordBindingFailure();
-                return null;
-            }
-
-            body = ApplyConversion(body, returnType, conversion, syntax.ExpressionBody);
-        }
-
-        lambdaBinder.SetLambdaBody(body);
-        var captured = lambdaBinder.AnalyzeCapturedVariables();
-
-        lambdaSymbol.SetCapturedVariables(captured);
-        lambdaSymbol.SetReturnType(returnType);
-        lambdaSymbol.SetDelegateType(delegateType);
-
-        var candidateDelegates = unbound.CandidateDelegates;
-        if (candidateDelegates.IsDefaultOrEmpty)
-        {
-            candidateDelegates = ImmutableArray.Create(delegateType);
-        }
-        else if (!candidateDelegates.Contains(delegateType, SymbolEqualityComparer.Default))
-        {
-            candidateDelegates = candidateDelegates.Add(delegateType);
-        }
-
-        _lambdaDelegateTargets[syntax] = candidateDelegates;
-
-        var rebound = new BoundLambdaExpression(
-            parameterSymbols,
-            returnType,
-            body,
-            lambdaSymbol,
-            delegateType,
-            captured,
-            candidateDelegates);
-        rebound.AttachUnbound(unbound);
-        instrumentation.RecordBindingSuccess();
-        return rebound;
-    }
-
-    private LambdaExpressionSyntax? GetLambdaSyntax(BoundLambdaExpression lambda)
-    {
-        if (lambda.Unbound is { Syntax: { } syntax })
-            return syntax;
-
-        if (lambda.Symbol is ILambdaSymbol lambdaSymbol)
-        {
-            foreach (var reference in lambdaSymbol.DeclaringSyntaxReferences)
-            {
-                if (reference.GetSyntax() is LambdaExpressionSyntax lambdaSyntax)
-                    return lambdaSyntax;
-            }
-        }
-
-        return null;
-    }
-
-    private void ReportSuppressedLambdaDiagnostics(IEnumerable<BoundExpression> arguments)
-    {
-        foreach (var argument in arguments)
-        {
-            if (argument is BoundLambdaExpression { Unbound: { } unbound })
-                unbound.ReportSuppressedDiagnostics(_diagnostics);
-        }
     }
 
     protected BoundExpression CreateOptionalArgument(IParameterSymbol parameter)
